+++ src/discord/commands/guild/link-accounts.js
@@ -4,7 +4,7 @@ const PREFIX = require('path').parse(__filename).name;
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { stripIndents } = require('common-tags/lib');
 const logger = require('../../../global/logger');
-const { getUserInfo, setUserInfo } = require('../../../global/firebase');
+const { getUserInfo, setUserInfo } = require('../../../global/firebaseAPI');
 const template = require('../../../global/embed-template');
 
 const {
