+++ src/discord/commands/guild/mod.js
@@ -7,7 +7,7 @@ const { MessageActionRow, MessageButton } = require('discord.js');
 const { stripIndents } = require('common-tags/lib');
 const logger = require('../../../global/logger');
 const template = require('../../../global/embed-template');
-const { getUserInfo, setUserInfo } = require('../../../global/firebase');
+const { getUserInfo, setUserInfo } = require('../../../global/firebaseAPI');
 const parseDuration = require('../../../global/parseDuration');
 
 const {
