+++ src/discord/commands/guild/remindme.js
@@ -6,7 +6,7 @@ const ms = require('ms');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const logger = require('../../../global/logger');
 const template = require('../../../global/embed-template');
-const { getUserInfo, setUserInfo } = require('../../../global/firebase');
+const { getUserInfo, setUserInfo } = require('../../../global/firebaseAPI');
 const parseDuration = require('../../../global/parseDuration');
 
 const PREFIX = path.parse(__filename).name;
