+++ src/discord/commands/guild/start-here.js
@@ -6,7 +6,7 @@ const { stripIndents } = require('common-tags');
 const { ReactionRole } = require('discordjs-reaction-role');
 const logger = require('../../../global/logger');
 const template = require('../../../global/embed-template');
-const { getGuildInfo, setGuildInfo } = require('../../../global/firebase');
+const { getGuildInfo, setGuildInfo } = require('../../../global/firebaseAPI');
 
 const {
   NODE_ENV,
