+++ src/discord/discordAPI.js
@@ -11,8 +11,8 @@ const {
 } = require('../../env');
 
 module.exports = {
-  async discordConnect() {
-    const client = new Client({
+  discordConnect: async () => {
+    const discordClient = new Client({
       intents: [
         Intents.FLAGS.GUILDS,
         Intents.FLAGS.GUILD_MESSAGES,
@@ -32,8 +32,11 @@ module.exports = {
         'REACTION',
       ],
     });
-    Promise.all([registerCommands(client), registerEvents(client)])
-      .then(() => client.login(discordToken))
+
+    Object.assign(global, { discordClient });
+
+    Promise.all([registerCommands(discordClient), registerEvents(discordClient)])
+      .then(() => discordClient.login(discordToken))
       .then(() => logger.info(`[${PREFIX}] Discord bot successfully started...`));
   },
 };
