+++ src/discord/events/guild-member-remove.js
@@ -3,7 +3,7 @@
 const PREFIX = require('path').parse(__filename).name;
 const logger = require('../../global/logger');
 const template = require('../../global/embed-template');
-const { getUserInfo } = require('../../global/firebase');
+const { getUserInfo } = require('../../global/firebaseAPI');
 
 const {
   discordGuildId,
