+++ src/global/experience.js
@@ -4,7 +4,7 @@ const path = require('path');
 
 const template = require('./embed-template');
 const logger = require('./logger');
-const { getUserInfo, setUserInfo } = require('./firebase');
+const { getUserInfo, setUserInfo } = require('./firebaseAPI');
 
 const PREFIX = path.parse(__filename).name;
