+++ src/global/firebaseAPI.js
@@ -1,18 +1,102 @@
 'use strict';
 
 const path = require('path');
+const fs = require('fs/promises');
+const { initializeApp, cert } = require('firebase-admin/app'); // eslint-disable-line
+const { getFirestore } = require('firebase-admin/firestore'); // eslint-disable-line
+const serviceAccount = require('../assets/config/firebase_creds.json');
+
 const logger = require('./logger');
 
 const PREFIX = path.parse(__filename).name;
 
-const { db } = global;
 const {
+  NODE_ENV,
   firebaseTicketDbName,
   firebaseGuildDbName,
   firebaseUserDbName,
+  firebasePrivateKeyId,
+  firebasePrivateKey,
+  firebaseClientId,
+  firebaseClientEmail,
 } = require('../../env');
 
+serviceAccount.private_key_id = firebasePrivateKeyId;
+serviceAccount.private_key = firebasePrivateKey ? firebasePrivateKey.replace(/\\n/g, '\n') : undefined;
+serviceAccount.client_email = firebaseClientEmail;
+serviceAccount.client_id = firebaseClientId;
+
+// Initialize firebase app
+if (serviceAccount.private_key_id) {
+  initializeApp({
+    credential: cert(serviceAccount),
+    databaseURL: 'https://tripsit-me-default-rtdb.firebaseio.com',
+  });
+  global.db = getFirestore();
+}
+
+const { db } = global;
+
 module.exports = {
+  firebaseConnect: async () => {
+    async function updateGlobalDb() {
+      const userDb = [];
+      if (db !== undefined) {
+        // Get user information
+        const snapshotUser = await db.collection(firebaseUserDbName).get();
+        snapshotUser.forEach(doc => {
+          userDb.push({
+            key: doc.id,
+            value: doc.data(),
+          });
+        });
+      }
+      Object.assign(global, { userDb });
+      // logger.debug(`[${PREFIX}] User database loaded.`);
+      // logger.debug(`[${PREFIX}] userDb: ${JSON.stringify(global.userDb, null, 4)}`);
+      logger.debug(`[${PREFIX}] Global User DB loaded!`);
+
+      const guildDb = [];
+      const blacklistGuilds = [];
+      if (db !== undefined) {
+        // Get guild information
+        const snapshotGuild = await db.collection(firebaseGuildDbName).get();
+        snapshotGuild.forEach(doc => {
+          guildDb.push({
+            key: doc.id,
+            value: doc.data(),
+          });
+          if (doc.data().isBanned) {
+            blacklistGuilds.push(doc.data().guild_id);
+          }
+        });
+      }
+      Object.assign(global, { guild_db: guildDb });
+      logger.debug(`[${PREFIX}] Global Guild DB loaded!`);
+      return userDb;
+    }
+
+    async function backupDb(userDb) {
+      const today = Math.floor(Date.now() / 1000);
+      if (NODE_ENV !== 'production') {
+        await fs.writeFile(
+          path.resolve(`./backups/userDb_(${today}).json`),
+          JSON.stringify(userDb, null, 2),
+        );
+        logger.debug(`[${PREFIX}] User database backedup!`);
+      }
+
+      if (NODE_ENV !== 'production') {
+        await fs.writeFile(
+          path.resolve(`./backups/guild_db_(${today}).json`),
+          JSON.stringify(global.guild_db, null, 2),
+        );
+        logger.debug(`[${PREFIX}] Guild database backedup!`);
+      }
+    }
+    const userDb = await updateGlobalDb();
+    await backupDb(userDb);
+  },
   getUserInfo: async member => {
     // logger.debug(`[${PREFIX}] member: ${JSON.stringify(member, null, 2)}`);
     // {
